Add usernames under avatars
As discussed in [this](https://community.nodebb.org/topic/2019/changing-template-file-with-a-plugin) topic, usernames should ideally be on the left, I have purposefully left the username class empty, so either we can come up with a solution to long usernames, or people can decide how they want it to appear, or not at all. 

You can see this in action [here](http://35hz.co.uk/topic/423/good-sci-fi-series). Feel free to improve on this if needed. :)